load_testing/requirements.txt
locust
PyYAML
